input: add keyboard symbols header and API to convert from/to unicode
Add scripts/update-keysyms-header to manually update the header when
needed. The script is based on the libxkbcommon and GDK ones, using
the installed X11 headers to create our own with the same symbols but
prefixed with WPE_KEY. The conversion to/from unicode is based on GDK
tables and conversion functions.
Now that we have functions to get the unicode for a given key code, we
no longer need to include the unicode as a member of the event struct.
We use that field now to include hardward key code.
wpe_input_key_mapper and wpe_input_key_mapper_interface are no longer
needed either, since the users can implement them by using the key
symbols and conversion to unicode API.