Rename the libWPEBackend library just to libwpe
This library does not actually implement any backend for WPE, and
therefore its name can be misleading. Instead, it contains the code
needed by WebKit to interface with actual backend implementations,
loading them, and some utility code. The libwpe name is much more
fitting.

This is a good moment to do this rename, because the API version will
be bumped for release 1.0.0 as well: packagers and users of the library
will need to update anyway for the API version change and rebuild
software which uses libWPEBackend^W libwpe anyway.

Fixes #27