cmake: Pass -Wall and check that compiler flags are supported
Adds checks to ensure that command line flags passed to the C/C++
copmilers are supported before adding them to the list of flags. This
also leaves to CMake the responsibility of determining which compiler
flags are needed to choose C++11/C99, instead of passing -std=xyz,
because different compilers might use other options for this.