cmake: Use an imported library for EGL
This modified FindEGL.cmake to create the GL::egl imported library
target, and use it in CMakeLists.txt. Note that only the EGL headers
are used (the library does not really need to be linked), so instead
of passing GL::egl to target_link_libraries() use generator expressions
to extract the needed include directories and compilation options from
the target.