Meson: Provide libxkbcommon subproject fallback
Add a .wrap subproject description in order to let Meson fetch and patch
libxkbcommon. The reason for the patching is that the libxkbcommon_dep
from upstream does not set the "include_directories" property of the
depedency object and that prevents using it as a subproject because the
top-level project does not get the flags needed for the compiler to find
the libxkbcommon headers. Also, import a .def file with the list of
exported functions, otherwise MSVC will fail to create an import
library.