Allow libwpe to be built statically
In the case where there will only ever be one WPE backend implemented for a platform the dynamic loading capabilities aren't needed. In this case allow the library to built statically.

By convention a WPE backend sets `_wpe_loader_interface` so that is used as the name for the `wpe_loader_interface`.

For CMake builds set `BUILD_SHARED_LIBS` to `OFF`. For Meson builds set `default_library` to `static`.