Explicitly enable xkb
The `xkbcommon` library is geared towards platforms using the XKB specification so it should be toggleable. Add a `WPE_ENABLE_XKB` option to compile xkb code. By default this option is on.

The contents of `wpe/input.h` that require `xkbcommon` are split into `wpe/input-xkb.h` The `src/input.c` was renamed to `src/input-xkb.c` to indicate the file is specific to xkb being enabled.

The pkg-config definition was expanded to support additional compilation flags and dependencies.