Format CMakeLists.txt file
Run cmake-format over all CMake files to get a consistent look and feel. A GitHub action was added which uses the cmake-format.py configuration. If there are any changes they are committed to the branch automatically.

Rename wpe.pc.cmake to wpe.pc.in so the formatter doesn't run over it.