Clarify CI workflows (#96)
Rename ci.yml to build.yml since it is explicitly building the project and documentation. Also make this workflow run on PR because whether the build was successful was not being beamed to a status check.

Rename pr.yml to format.yml since it is explicitly ensuring that the source code is properly formatted.