Fix wrong license header in libwpe-version.h
Change the source comment at the top of libwpe-version.h to contain the
blurb of the BSD-2-Clause license, which is the intended one. The line
about that particular source being MIT was a mistake which ought to be
corrected.

(cherry picked from commit 7bdecb90e77c71a9d949a93dbd7007f4dce60f5a)