Create GL::egl target without pkgconfig
The if statement used `EGL_LIBRARIES` which was never created when not using the pkgconfig path so the `GL::egl` target wasn't there. Modify the check to use `EGL_LIBRARY`.

Add the `FOUND_VAR` to the `find_package_handle_standar_args` and add a way to set the name of the EGL implementation.